# Introducing python setup.py rebuild develop