@ezyang 2018-09-10
> Note: rebuild should only be used if you dont add / remove files to the build, as cmake is not rerun

For any non-glob file lists, cmake will automatically rerun itself on a change ;)